Added clarification around presentation only
Make sure it is clear to everyone that it is not the purpose here to create the model for the billing system (that calculates all the charges). We can assume all those charges (e.g. the price I paid when I sent an SMS, etc.) has already been calculated and persisted for us.